Multi-source-ETL-Dashboard-main/multi_source_etl_dashboard/requirements.txt
pandas 
streamlit 
snowflake-connector-python 
apache-airflow 
requests 